refactoring: move classes: filter.Record -> object.Record, filter.JSONException -> object.JSONParseException
