+++ src/main/java/org/bricolages/streaming/filter/DeleteNullsOp.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming.filter;
+import org.bricolages.streaming.object.Record;
 import java.util.Map;
 import lombok.*;
