+++ src/main/java/org/bricolages/streaming/filter/IntOp.java
@@ -1,5 +1,6 @@
 package org.bricolages.streaming.filter;
 import org.bricolages.streaming.stream.processor.Cleanse;
+import org.bricolages.streaming.object.Record;
 import lombok.*;
 
 class IntOp extends SingleColumnOp {
