+++ src/main/java/org/bricolages/streaming/filter/RenameOp.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming.filter;
+import org.bricolages.streaming.object.Record;
 import lombok.*;
 
 public class RenameOp extends Op {
