+++ src/main/java/org/bricolages/streaming/filter/SequenceOp.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming.filter;
+import org.bricolages.streaming.object.Record;
 import org.bricolages.streaming.exception.ApplicationError;
 import lombok.*;
