+++ src/main/java/org/bricolages/streaming/filter/TimeZoneOp.java
@@ -1,5 +1,6 @@
 package org.bricolages.streaming.filter;
 import org.bricolages.streaming.stream.processor.Cleanse;
+import org.bricolages.streaming.object.Record;
 import java.time.ZoneOffset;
 import lombok.*;
