+++ src/test/java/org/bricolages/streaming/stream/processor/StringColumnProcessorTest.java
@@ -1,6 +1,6 @@
 package org.bricolages.streaming.stream.processor;
 import org.bricolages.streaming.stream.StreamColumn;
-import org.bricolages.streaming.filter.Record;
+import org.bricolages.streaming.object.Record;
 import org.bricolages.streaming.filter.FilterException;
 import org.bricolages.streaming.exception.*;
 import java.util.ArrayList;
